Funções de cadastro mandadas para util
